+++ test/StrictPinSpec.hs
@@ -53,6 +53,7 @@ import Types (
     LocationSource (..),
     SimpleDatabase (..),
     SparseTriple (..),
+    SupplierClaim (..),
     TechRole (..),
     TechnosphereFlow (..),
     UUID,
@@ -249,7 +250,7 @@ supplierDB offset products =
                         , techUnitId = kgUnitId
                         , techRole = ReferenceProduct
                         , techActivityLinkId = actUUID
-                        , techSupplierActivity = Nothing
+                        , techSupplierClaim = ClaimByProduct
                         , techLocation = ""
                         , techComment = Nothing
                         , techPedigree = Nothing
@@ -302,7 +303,7 @@ consumerDB offset products =
                         , techUnitId = kgUnitId
                         , techRole = ReferenceProduct
                         , techActivityLinkId = actUUID
-                        , techSupplierActivity = Nothing
+                        , techSupplierClaim = ClaimByProduct
                         , techLocation = ""
                         , techComment = Nothing
                         , techPedigree = Nothing
@@ -317,7 +318,7 @@ consumerDB offset products =
                         , techUnitId = kgUnitId
                         , techRole = Input
                         , techActivityLinkId = UUID.nil
-                        , techSupplierActivity = Nothing
+                        , techSupplierClaim = ClaimByProduct
                         , techLocation = "GLO"
                         , techComment = Nothing
                         , techPedigree = Nothing